requirements.txt
pymupdf
pdfplumber
pytesseract
Pillow
nltk
scikit-learn
networkx
google-generativeai
streamlit
transformers
torch
flask
rouge-score
psutil
